+++ src/CommandLine/Names/OptionNames.cs
@@ -28,12 +28,10 @@ internal static class OptionNames
     public const string Context = "context";
     public const string Count = "count";
     public const string DetectRename = "detect-rename";
-    public const string Display = "display";
     public const string DryRun = "dry-run";
     public const string Encoding = "encoding";
     public const string EndsWith = "ends-with";
     new public const string Equals = "equals";
-    public const string Evaluator = "evaluator";
     public const string ExplicitCapture = "explicit-capture";
     public const string Extension = "extension";
     public const string Filter = "filter";